Extract request-response-related code to separate module
